pyproject.toml
[project]
name = "digital-badge-app"
version = "0.1.0"
description = "Digital badge generation and sharing platform"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
    "pillow>=10.0.0",
    "qrcode>=7.4.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
